Fix 64-bit NMInt32 values. Remove op_dlfcn: use built-in dlfcn which has been part of macOS since 10.3 (IIRC). Use CF_ENUMs/CF_OPTIONs.

+++ Source/Demos/OPDownloadHTTP/OPDownloadHTTP.cpp
@@ -117,7 +117,7 @@ int main(int argc, char **argv)
 	{
 		OPHTTPDownload *theDownload;
 		char buffer[5];
-		unsigned long bufferSize;
+		unsigned int bufferSize;
 		OPHTTPDownloadStatus dlStatus;
 		unsigned int bytesReceived = 0;
